made changes for classes/MainTask.java, fundamentals/Task1

+++ src/main/java/course/fundamentals/OptionalTask1.java
@@ -100,10 +100,10 @@ public class OptionalTask1 {
         System.out.println("Average length: " + average);
 
         for (int i = 0; i < array.length; i++) {
-            if (String.valueOf(array[i]).length() < average) {
-                System.out.println("has less length than average: " + array[i] + " " + "\n" + "length: " + String.valueOf(array[i]).length());
+            if ((double) String.valueOf(array[i]).length() < average) {
+                System.out.println("has less length than average: " + array[i] + " " + "\n" + "length: " + (double) String.valueOf(array[i]).length());
             } else {
-                System.out.println("has length more than average: " + array[i] + " " + "\n" + "length: " + String.valueOf(array[i]).length());
+                System.out.println("has length more than average: " + array[i] + " " + "\n" + "length: " + (double) String.valueOf(array[i]).length());
             }
         }
     }
